refactor: update asset references in index.html and replace obsolete JavaScript file
Updated the `index.html` file to reference a new JavaScript asset, replacing the obsolete `index-CRLnkQAO.js` with `index-XS_LOCUO.js`. This change enhances the application's performance and ensures the latest features are utilized. The previous JavaScript file has been removed from the `dist` directory, streamlining the build process.

Added experimentStartedAt and experimentFinishedAt in pdf as well as in DB